This commit resolves issue #116 "ClassCastException when setting integer value for σ in noise-data component."
Reworked the handling of DSL files for developers. The `LanguageHelper` component was renamed to `AttributeHelper` and is more powerful than the old version. The function `lookup` automatically does default value handling (using default values of the attribute definition if the attribute is missing in the DSL) and automatic type casts. Thus, `int` values are now automatically casted into `double` values if the attribute definition states that the attribute is of type `real`.

Changelog: changed